refactor(#652): move unsubscribe, prices page to app router
