+++ frontend/src/app/(scoreboard)/dojodigest/unsubscribe/page.tsx
@@ -1,0 +1,5 @@
+import UnsubscribePage from './UnsubscribePage';
+
+export default function Page() {
+    return <UnsubscribePage />;
+}
