+++ frontend/src/app/(scoreboard)/prices/page.tsx
@@ -1,0 +1,10 @@
+import { Suspense } from 'react';
+import PricingPage from './PricingPage';
+
+export default function Page() {
+    return (
+        <Suspense>
+            <PricingPage />
+        </Suspense>
+    );
+}
