+++ frontend/src/profile/creator/ProfileCreatorPage.tsx
@@ -1,8 +1,8 @@
 import { Box, Container, Step, StepLabel, Stepper, Typography } from '@mui/material';
 import { useState } from 'react';
+import PricingPage from '../../app/(scoreboard)/prices/PricingPage';
 import { useRequiredAuth } from '../../auth/Auth';
 import { SubscriptionStatus, User, dojoCohorts } from '../../database/user';
-import PricingPage from '../../upsell/PricingPage';
 import DiscordForm from './DiscordForm';
 import ExtraRatingSystemsForm from './ExtraRatingSystemsForm';
 import PersonalInfoForm from './PersonalInfoForm';
